feat: revert ratings back
